clean up bool array module
