+++ vortex-array/src/arrays/bool/vtable/array.rs
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_dtype::DType;
+
+use crate::arrays::{BoolArray, BoolVTable};
+use crate::stats::StatsSetRef;
+use crate::vtable::ArrayVTable;
+
+impl ArrayVTable<BoolVTable> for BoolVTable {
+    fn len(array: &BoolArray) -> usize {
+        array.buffer.len()
+    }
+
+    fn dtype(array: &BoolArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &BoolArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+}
