+++ vortex-array/src/arrays/bool/vtable/operations.rs
