+++ vortex-array/src/arrays/bool/vtable/serde.rs
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-use vortex_buffer::{Alignment, ByteBuffer};
+use vortex_buffer::ByteBuffer;
 use vortex_dtype::DType;
 use vortex_error::{VortexResult, vortex_bail, vortex_err};
 
 use super::BoolArray;
+use crate::ProstMetadata;
 use crate::arrays::BoolVTable;
 use crate::serde::ArrayChildren;
 use crate::validity::Validity;
-use crate::vtable::{SerdeVTable, VTable, VisitorVTable};
-use crate::{ArrayBufferVisitor, ArrayChildVisitor, ProstMetadata};
+use crate::vtable::{SerdeVTable, VTable};
 
 #[derive(prost::Message)]
 pub struct BoolMetadata {
@@ -51,18 +51,5 @@ impl SerdeVTable<BoolVTable> for BoolVTable {
         };
 
         BoolArray::try_new(buffers[0].clone(), metadata.offset as usize, len, validity)
-    }
-}
-
-impl VisitorVTable<BoolVTable> for BoolVTable {
-    fn visit_buffers(array: &BoolArray, visitor: &mut dyn ArrayBufferVisitor) {
-        visitor.visit_buffer(&ByteBuffer::from_arrow_buffer(
-            array.boolean_buffer().clone().into_inner(),
-            Alignment::none(),
-        ))
-    }
-
-    fn visit_children(array: &BoolArray, visitor: &mut dyn ArrayChildVisitor) {
-        visitor.visit_validity(&array.validity, array.len());
     }
 }
